refactor: resolve Codacy complexity and file-length findings
Extract deleteEntry and copyEntry helpers in the files store, move ZIP
member validation out of inspect_zip, and split the MCP pending-call
tests into their own module so every function and file sits under the
Lizard limits.
